+++ src/xgboost_distribution/distributions/__init__.py
@@ -28,7 +28,7 @@ AVAILABLE_DISTRIBUTIONS = {
 def get_distribution(name):
     """Get instantianted distribution based on name"""
 
-    if name not in AVAILABLE_DISTRIBUTIONS.keys():
+    if name not in AVAILABLE_DISTRIBUTIONS:
         raise ValueError(
             "Distribution is not implemented! Please choose one of "
             f"{set(AVAILABLE_DISTRIBUTIONS.keys())}"
@@ -52,6 +52,9 @@ def get_distribution_doc():
     param_doc += """
         Please see `scipy.stats` for a full description of the parameters of
         each distribution: https://docs.scipy.org/doc/scipy/reference/stats.html
+
+        Note that distributions are fit using Maximum Likelihood Estimation, which
+        internally corresponds to minimising the negative log likelihood (NLL).
     """
 
     return param_doc