+++ src/xgboost_distribution/distributions/base.py
@@ -1,7 +1,9 @@
-"""Distribution base class
-"""
+"""Distribution base class"""
+
 from abc import ABC, abstractmethod
-from collections import namedtuple
+from typing import Tuple
+
+import numpy as np
 
 
 class BaseDistribution(ABC):
@@ -11,31 +13,28 @@ class BaseDistribution(ABC):
     functions that operate on the data (`y`) and the outputs of the xgboost (`params`).
     """
 
-    def __init__(self):
-        self.Predictions = namedtuple("Predictions", (p for p in self.params))
-        # attach to globals to make pickling of namedtuple work
-        globals()[self.Predictions.__name__] = self.Predictions
-
-    def check_target(self, y):
+    def check_target(self, y) -> None:
         """Ensure that the target is compatible with the chosen distribution"""
 
     @property
     @abstractmethod
-    def params(self):
+    def params(self) -> Tuple[str, ...]:
         """The parameter names of the distribution"""
 
     @abstractmethod
-    def starting_params(self, y):
+    def starting_params(self, y) -> Tuple[float, ...]:
         """The starting parameters of the distribution"""
 
     @abstractmethod
-    def gradient_and_hessian(self, y, params, natural_gradient=True):
+    def gradient_and_hessian(
+        self, y, params, natural_gradient=True
+    ) -> Tuple[np.ndarray, np.ndarray]:
         """Compute the gradient and hessian of the distribution"""
 
     @abstractmethod
-    def loss(self, y, params):
-        """Evaluate the loss (typically negative log-likelihood)"""
+    def loss(self, y, params) -> Tuple[str, np.ndarray]:
+        """Evaluate the per sample loss (typically negative log-likelihood)"""
 
     @abstractmethod
-    def predict(self, params):
+    def predict(self, params) -> Tuple[np.ndarray, ...]:
         """Predict the parameters of a given distribution"""