+++ src/xgboost_distribution/distributions/negative_binomial.py
@@ -1,11 +1,22 @@
-"""Negative binomial distribution
-"""
+"""Negative binomial distribution"""
+
+from collections import namedtuple
+
 import numpy as np
 from scipy.special import digamma, expit
 from scipy.stats import nbinom
 
+from xgboost_distribution.compat import linalg_solve
 from xgboost_distribution.distributions.base import BaseDistribution
-from xgboost_distribution.distributions.utils import check_is_ge_zero, check_is_integer
+from xgboost_distribution.distributions.utils import (
+    MAX_EXPONENT,
+    MIN_EXPONENT,
+    check_all_ge_zero,
+    check_all_integer,
+    safe_exp,
+)
+
+Params = namedtuple("Params", ("n", "p"))
 
 
 class NegativeBinomial(BaseDistribution):
@@ -65,33 +76,29 @@ class NegativeBinomial(BaseDistribution):
 
     @property
     def params(self):
-        return ("n", "p")
+        return Params._fields
 
     def check_target(self, y):
-        check_is_integer(y)
-        check_is_ge_zero(y)
+        check_all_integer(y)
+        check_all_ge_zero(y)
 
     def gradient_and_hessian(self, y, params, natural_gradient=True):
         """Gradient and diagonal hessian"""
 
-        log_n, raw_p = params[:, 0], params[:, 1]
-        n = np.exp(log_n)
-        p = expit(raw_p)
+        n, p = self.predict(params)
 
-        grad = np.zeros(shape=(len(y), 2))
+        grad = np.zeros(shape=(len(y), 2), dtype="float32")
 
         grad[:, 0] = -n * (digamma(y + n) - digamma(n) + np.log(p))
         grad[:, 1] = p * (y - n * (1 - p) / p)
 
         if natural_gradient:
-
-            fisher_matrix = np.zeros(shape=(len(y), 2, 2))
+            fisher_matrix = np.zeros(shape=(len(y), 2, 2), dtype="float32")
             fisher_matrix[:, 0, 0] = (n * p) / (p + 1)
             fisher_matrix[:, 1, 1] = n * p
 
-            grad = np.linalg.solve(fisher_matrix, grad)
-            hess = np.ones(shape=(len(y), 2))  # we set the hessian constant
-
+            grad = linalg_solve(fisher_matrix, grad)
+            hess = np.ones(shape=(len(y), 2), dtype="float32")  # constant hessian
         else:
             raise NotImplementedError(
                 "Normal gradients are currently not supported by this "
@@ -102,14 +109,17 @@ class NegativeBinomial(BaseDistribution):
 
     def loss(self, y, params):
         n, p = self.predict(params)
-        return "NegativeBinomial-NLL", -nbinom.logpmf(y, n=n, p=p).mean()
+        return "NegativeBinomial-NLL", -nbinom.logpmf(y, n=n, p=p)
 
     def predict(self, params):
-        log_n, raw_p = params[:, 0], params[:, 1]
-        n = np.exp(log_n)
-        p = expit(raw_p)
-        return self.Predictions(n=n, p=p)
+        log_n, logits = params[:, 0], params[:, 1]
+
+        n = safe_exp(log_n)
+        logits = np.clip(logits, a_min=MIN_EXPONENT, a_max=MAX_EXPONENT)
+
+        p = expit(logits)
+        return Params(n=n, p=p)
 
     def starting_params(self, y):
         # TODO: starting params can matter a lot?
-        return (np.log(np.mean(y)), 0)  # expit(0) = 0.5
+        return Params(n=np.log(np.mean(y)), p=0)  # expit(0) = 0.5