+++ src/xgboost_distribution/distributions/poisson.py
@@ -1,10 +1,19 @@
-"""Poisson distribution
-"""
+"""Poisson distribution"""
+
+from collections import namedtuple
+
 import numpy as np
 from scipy.stats import poisson
 
+from xgboost_distribution.compat import linalg_solve
 from xgboost_distribution.distributions.base import BaseDistribution
-from xgboost_distribution.distributions.utils import check_is_ge_zero, check_is_integer
+from xgboost_distribution.distributions.utils import (
+    check_all_ge_zero,
+    check_all_integer,
+    safe_exp,
+)
+
+Params = namedtuple("Params", ("mu"))
 
 
 class Poisson(BaseDistribution):
@@ -29,40 +38,38 @@ class Poisson(BaseDistribution):
 
     @property
     def params(self):
-        return ("mu",)
+        return Params._fields
 
     def check_target(self, y):
-        check_is_integer(y)
-        check_is_ge_zero(y)
+        check_all_integer(y)
+        check_all_ge_zero(y)
 
     def gradient_and_hessian(self, y, params, natural_gradient=True):
         """Gradient and diagonal hessian"""
 
         (mu,) = self.predict(params)
 
-        grad = np.zeros(shape=(len(y), 1))
+        grad = np.zeros(shape=(len(y), 1), dtype="float32")
         grad[:, 0] = mu - y
 
         if natural_gradient:
-            fisher_matrix = np.zeros(shape=(len(y), 1, 1))
+            fisher_matrix = np.zeros(shape=(len(y), 1, 1), dtype="float32")
             fisher_matrix[:, 0, 0] = mu
 
-            grad = np.linalg.solve(fisher_matrix, grad)
-
-            hess = np.ones(shape=(len(y), 1))  # we set the hessian constant
+            grad = linalg_solve(fisher_matrix, grad)
+            hess = np.ones(shape=(len(y), 1), dtype="float32")  # constant hessian
         else:
             hess = mu
 
         return grad, hess
 
     def loss(self, y, params):
-        mu = self.predict(params)
-        return "Poisson-NLL", -poisson.logpmf(y, mu=mu).mean()
+        (mu,) = self.predict(params)
+        return "Poisson-NLL", -poisson.logpmf(y, mu=mu)
 
     def predict(self, params):
-        log_mu = params
-        mu = np.exp(log_mu)
-        return self.Predictions(mu=mu)
+        mu = safe_exp(params)
+        return Params(mu=mu)
 
     def starting_params(self, y):
-        return (np.log(np.mean(y)),)
+        return Params(mu=np.log(np.mean(y)))