+++ tests/distributions/test_exponential.py
@@ -43,17 +43,38 @@ def test_target_validation_raises(exponential, invalid_target):
     ],
 )
 def test_gradient_calculation(exponential, y, params, natural_gradient, expected_grad):
-    grad, hess = exponential.gradient_and_hessian(
+    grad, _ = exponential.gradient_and_hessian(
         y, params, natural_gradient=natural_gradient
     )
     np.testing.assert_array_equal(grad, expected_grad)
 
 
 def test_loss(exponential):
-    loss_name, loss_value = exponential.loss(
-        # fmt: off
-        y=np.array([1, ]),
-        params=np.array([[np.log(1)], ]),
+    loss_name, loss_values = exponential.loss(
+        y=np.array([1]),
+        params=np.array([np.log(1)]),
     )
     assert loss_name == "Exponential-NLL"
-    assert loss_value == 1.0
+    np.testing.assert_array_equal(loss_values, np.array([1.0]))
+
+
+@pytest.mark.parametrize(
+    "y, params",
+    [
+        (
+            np.array([34.7, 20.1], dtype="float32"),
+            np.array([150, 20], dtype="float32"),
+        ),
+        (
+            np.array([34.7, 20.1], dtype="float32"),
+            np.array([-150, 20], dtype="float32"),
+        ),
+    ],
+)
+def test_overflow_stability(exponential, y, params):
+    """Test stability against large/small values produced by xgboost"""
+
+    # Scale should not have inf or zeros
+    scale = exponential.predict(params)
+    assert np.isfinite(scale).all()
+    assert np.all(scale)