+++ CHANGELOG.rst
@@ -5,13 +5,72 @@ Changelog
 Development version
 ===================
 
-Version 0.2.X, 2022-xx-xx
+Version 0.3.x, 20xx-xx-xx
 --------------------------
 
-- ...
+- ..
+
 
 Current version
 ===============
+
+Version 0.3.1, 2025-02-15
+--------------------------
+
+- Fix for sklearn >= 1.6 compatibility, :issue:`104`
+
+
+Older versions
+===============
+
+
+Version 0.3.0, 2024-07-06
+--------------------------
+
+- Adapted for xgboost==2.1.0 compatibility
+- Numpy compatibility for beyond 2.0.0 (older versions are still supported)
+
+
+Version 0.2.9, 2023-12-10
+--------------------------
+
+- Fix float32 serialisation, :issue:`97`
+
+
+Version 0.2.8, 2023-09-24
+--------------------------
+
+- Fixes to make package work with xgboost 2.0
+- Various internal changes for xgboost sklearn API consistency
+
+
+Version 0.2.7, 2023-03-12
+--------------------------
+
+- Fix overflow issues for normal distribution, :issue:`64`
+- Removed verbosity hack in model training
+- Better support for pickle/joblib, :issue:`82`
+
+
+Version 0.2.6, 2023-01-21
+--------------------------
+
+- Added support for sample weights, :issue:`45`
+
+
+Version 0.2.5, 2022-11-01
+--------------------------
+
+- Added example script for hyperparameter tuning
+- Python requires >= 3.8 & xgboost >= 1.7.0 compatibility
+
+
+Version 0.2.4, 2022-04-23
+--------------------------
+
+- Added more precise loss description, negative log likelihood vs error
+- Various updates to conform with xgboost==1.6.0 release
+
 
 Version 0.2.3, 2021-12-28
 --------------------------
@@ -19,9 +78,6 @@ Version 0.2.3, 2021-12-28
 - Added type hints to XGBDistribution model class
 - Hotfix to add error raising if sample weights are used (which is not yet implemented)
 
-
-Older versions
-===============
 
 Version 0.2.2, 2021-10-23
 --------------------------